requirements.txt
playwright>=1.48.0
beautifulsoup4>=4.12.2
lxml>=5.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.9.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
openai>=1.0.0
google-genai>=1.0.0

